Correction of methods

+++ src/test/java/ru/netology/stats/StatsServiceTest.java
@@ -8,54 +8,54 @@ public class StatsServiceTest {
     @Test
     public void shouldSummarySales() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expectedSum = 180;
-        int actualSum = service.salesAmount(sales);
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long expectedSum = 180;
+        long actualSum = service.salesAmount(sales);
         Assertions.assertEquals(expectedSum, actualSum);
     }
 
     @Test
     public void shouldAverageSales() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expectedSum = 15;
-        int actualSum = service.salesAverage(sales);
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long expectedSum = 15;
+        long actualSum = service.salesAverage(sales);
         Assertions.assertEquals(expectedSum, actualSum);
     }
 
     @Test
     public void shouldMaxMonthSales() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expectedSum = 8;
-        int actualSum = service.maxSales(sales);
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long expectedSum = 8;
+        long actualSum = service.maxSales(sales);
         Assertions.assertEquals(expectedSum, actualSum);
     }
 
     @Test
     public void shouldMixMonthSales() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expectedSum = 9;
-        int actualSum = service.minSales(sales);
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long expectedSum = 9;
+        long actualSum = service.minSales(sales);
         Assertions.assertEquals(expectedSum, actualSum);
     }
 
     @Test
     public void shouldMinAvgMonthSales() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expectedSum = 5;
-        int actualSum = service.minAvgSales(sales);
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long expectedSum = 5;
+        long actualSum = service.minAvgSales(sales);
         Assertions.assertEquals(expectedSum, actualSum);
     }
 
     @Test
     public void shouldMaxAvgMonthSales() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expectedSum = 5;
-        int actualSum = service.maxAvgSales(sales);
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long expectedSum = 5;
+        long actualSum = service.maxAvgSales(sales);
         Assertions.assertEquals(expectedSum, actualSum);
     }
 }
